+++ src/index.ts
@@ -8,7 +8,13 @@ export {
 export * from './interfaces.js';
 export { createDID, deactivateDID, resolveDID, resolveDIDFromLog, updateDID } from './method.js';
 export { MultibaseEncoding, multibaseDecode, multibaseEncode } from './utils/multiformats.js';
-export { generateParallelDidWeb, parseDidKeyDid, parseDidKeyVerificationMethod } from './utils.js';
+export {
+  createVMID,
+  DID_PLACEHOLDER,
+  generateParallelDidWeb,
+  parseDidKeyDid,
+  parseDidKeyVerificationMethod,
+} from './utils.js';
 export {
   createWitnessProof,
   signWitnessProofEntries,